+++ plugin-packs/postcss-preset-env/src/lib/shared-options.mjs
@@ -1,11 +1,7 @@
 export function initializeSharedOptions(opts) {
 	if ('preserve' in opts) {
 		const sharedOptions = {};
-
-		if ('preserve' in opts) {
-			sharedOptions.preserve = opts.preserve;
-		}
-
+		sharedOptions.preserve = opts.preserve;
 		return sharedOptions;
 	}