+++ plugins/postcss-logical/src/utils/logical-to-physical.ts
@@ -1,0 +1,20 @@
+import { DirectionConfig, DirectionValues } from '../lib/types';
+
+export function logicalToPhysical(logical: string, config: DirectionConfig): string {
+	const [blockStart, blockEnd] = config.block;
+	const [inlineStart, inlineEnd] = config.inline;
+
+	switch (logical) {
+		case DirectionValues.BlockStart:
+			return blockStart;
+
+		case DirectionValues.BlockEnd:
+			return blockEnd;
+
+		case DirectionValues.InlineStart:
+			return inlineStart;
+
+		case DirectionValues.InlineEnd:
+			return inlineEnd;
+	}
+}